+++ docs/make.jl
@@ -5,8 +5,12 @@ using RiskPerf
 
 makedocs(;
     sitename="RiskPerf.jl",
-    # format = Documenter.HTML(prettyurls=false),
-    # modules = [RiskPerf]
+    format=Documenter.HTML(;
+        prettyurls=get(ENV, "CI", nothing) == "true",
+        sidebar_sitename=false,
+        assets=["assets/styles.css"],
+        edit_link="main",
+    ),
 )
 
-deploydocs(; repo="github.com/rbeeli/RiskPerf.jl.git")
+deploydocs(; repo="github.com/rbeeli/RiskPerf.jl.git", devbranch="main")