+++ test/information_ratio.jl
@@ -1,0 +1,3 @@
+@testitem "information_ratio" setup = [TestData] begin
+    @test information_ratio(asset_returns, market_returns) ≈ 0.34034622188359
+end